GUI: Can start/stop/pause training and change training options

+++ Retia/Training/Trainers/OptimizationReportEventArgs.cs
@@ -4,13 +4,13 @@ namespace Retia.Training.Trainers
 {
 	public class OptimizationReportEventArgs : TrainReportEventArgsBase
 	{
-		public OptimizationReportEventArgs(List<double> errors, double epoch, long iteration, double learningRate) : base(epoch, iteration)
+		public OptimizationReportEventArgs(List<double> errors, double epoch, long iteration, float learningRate) : base(epoch, iteration)
 		{
 			Errors = errors;
 			LearningRate = learningRate;
 		}
 
 		public List<double> Errors { get; set; }
-        public double LearningRate { get; set; }
+        public float LearningRate { get; set; }
 	}
 }
